separates between builtin modules and plugin modules
