+++ python/msa/builtins/command/event.py
@@ -1,0 +1,3 @@
+import msa.event
+
+
