+++ python/msa/builtins/command/module.py
@@ -1,0 +1,12 @@
+
+from msa.plugins.conversation import event
+from msa.plugins.conversation import coroutine
+
+class PluginModule:
+
+    coroutines = [
+        coroutine.ConversationCoroutine()
+    ]
+
+    events = []
+
